Slow marquee name.
Trying to go faster causes rate limiting. It seems that the slack api
allows short bursts of sub-second calls, but doesn't like the average
time between calls to be much less than about 7 seconds or so. This is
dumb, but not much I can do about it. Their "Real Time Messaging" API
appears to allow for editing user data despite the name, so I might try
that.